This has abilities impleneted correctly
Use the end turn button to activate abilities, it will at some point also be able to use the battle script but for now it just does the abilities. press on one instance of the game and wait for "ETB Triggers" to appear on the button to see the effect. This sould make adding more card effects easy to do.